refactor(log | history): 중복되는 ListHeader 컴포넌트 통합 (#37)
- 기존에 분리되어 있던 ListHeader 컴포넌트를 하나로 병합
- headerTitles props로 헤더 제목 리스트를 받도록 개선
- 컴포넌트 재사용성 향상 및 코드 중복 제거
------------------------------------------------------

------------------------------------------------------
